// src/app/modules/configuration/general/models/school-virtual-classes-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * SchoolVirtualClassesView
 */
export interface SchoolVirtualClassesView { 
    /**
     * Gets or Sets ClassId
     */
    classId?: number;
    /**
     * Gets or Sets SectionId
     */
    sectionId?: number;
    /**
     * Gets or Sets CourseId
     */
    courseId?: number;
    /**
     * Gets or Sets VirtualClassConfigurationId
     */
    virtualClassConfigurationId?: number;
    /**
     * Gets or Sets Virtual_Class_Info
     */
    virtualClassInfo?: string;
    /**
     * Gets or Sets IsActive
     */
    isActive?: number;
}