// src/app/modules/configuration/general/models/school-virtual-class-status-view.ts
/**
 * BaseWebAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


/**
 * Gets or Sets VirtualClassStatusView
 */
export interface SchoolVirtualClassStatusView { 
    /**
     * Gets or Sets IsActive
     */
    isActive?: boolean;
    id?: string;
}